bit of refactoring to clean up and reuse and work towards reusable autocorrect
